fix: typos all over the shot and added a couple of nanotrade discord quotes (#63)
* Fixing typo

Why do I have to do all the hard stuff??

* Fixing typo

Another typo to fix... My back hurts

* Fixing typo

* Fixing typo

Should Nano always be capitalized? I don't know...

* Fixing typo

I find the use of "can not" offensive.

* Fixing typo

Oh come on

* Fixing typo

* Fixing typo

For this one I had to fix the typo and then remove a space to pull the pipe back over. This was a tough one.

* Fixing typo

* Added a couple winners